README.md
This project is a clean, responsive portfolio website developed in HTML and CSS, showcasing my programming skills, educational background, and personal projects. It includes a downloadable CV, contact form, and links to GitHub and LinkedIn. Built as part of my journey into software development and participation in hackathons, this site is fully hosted using GitHub Pages.